Eq instance exercise & ghcid
Remove a function that if implemented could easily cause an import
cycle.

Add an exercise for implementing the `Eq` instance for `MyBTree` and
then checking that it satisfies the laws.

Leaving the completed exercise in for the moment to check my working.

Updated the readme to include some instructions for configuring ghcid to
run tests when the code is buildable. Needs finessing as there are some
wrinkles that won't be obvious.