Add hedgehog-fn for Eq instance tests
Added hedgehog-fn for generating the (a -> Bool) function for checking
the substitutivity property. Not sure if this pushes the weirdness
needle too far as we can't use 'Char' in the value position as there
is no Generic instance for it.

Added some more notes to push people on to the other exercises and try
to keep people focused when writing the Eq instance.